hammingCoder: translate from C to C++20, lots of stuff
C++20 has many useful features that enabled me to change (or just
remove) my code so it would be nicer.

Now I use std::vector instead of fussing with realloc etc.

Also included are some slight changes to the data structure to make it
alignable to a user provided number of bytes, to facilitate automatic
vectorizing by the compiler. Alas, the compiler seems to ignore my
efforts currently.

The C++20 functions from <bit> replace some of my bit twiddling
functions.

The fuzzing now checks the Hamming code coding against a naively
implemented coder.

nofuzz.cc and the build targets with "nofuzz" in their names are meant
for debugging after a fuzzer finds a bug. In a real project they could
also be used for regression testing. The targets with "gdb" in their
names are meant for use with a debugger tool.