hammingCoder: add new algos: Rows, RowsSparse
Rows is based on an imagined generator matrix approach, like Cols, but
approaches the vector-matrix multiplication by iterating on the rows of
the matrix (instead of columns) in the outer loop.

RowsSparse is based on the same idea, but works with an in-memory
precomputed sparse generator matrix.

Both work with fat bit vectors (one bit per char).

Fastest algorithms depending on the code are currently:

For n=7:    RowsDense.

For n=63:   RowsSparse.

For n=127:  RowsDense.

For n=1023: a tie between RowsDense and RowsSparse.